backend/requirements.txt
fastapi==0.104.1
python-socketio==5.10.0
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
pytest==7.4.3
hypothesis==6.92.1
httpx==0.25.2
requests==2.31.0
pyaudio==0.2.14
numpy==1.24.3
pywin32==306; sys_platform == 'win32'
